\name{dilution.phenodata}
\docType{data}
\alias{dilution.phenodata}
\title{Phenotypic Information for Dilution Study}
\description{
This objact is of class \code{\link[Biobase]{phenoData}} with necessary
  information for the assessemnts.
}
\usage{data(dilution.phenodata)}
\format{An object of class \code{\link[Biobase]{phenoData}}}
\source{Two sources of cRNA A (human liver tissue) and B (Central Nervous
System cell line) have been hybridized to human array (HGU95Av2) in a range
of proportions and dilutions. This object described these.

For more information see Irizarry, R.A., et al. (2001)
\url{http://www.biostat.jhsph.edu/~ririzarr/papers/index.html}}


\keyword{datasets}
