\name{hgu133a.spikein.phenodata}
\docType{data}
\alias{hgu133a.spikein.phenodata}
\title{phenotypic information for HGU133A spike in study}
\description{
This objact is of class \code{\link[Biobase]{phenoData}} with necessary
  information for the assessemnts.
}
\usage{data(hgu133a.spikein.phenodata)}
\format{An object of class \code{\link[Biobase]{phenoData}}}
\source{This comes from an experiments where 16 different cRNA
fragments have been added to the hybridization mixture of the GeneChip
arrays at different pM concentrations.  
For more information see Irizarry, R.A., et al. (2001)
\url{http://www.biostat.jhsph.edu/~ririzarr/papers/index.html}}


\keyword{datasets}
