\name{lw.sd.assessment}
\docType{instance}
\alias{lw.sd.assessment}
\title{An example of the result of an SD assessment}
\description{
  The Dilution files were processed with the
  dChip package (using PM-only) and then the
  function \code{\link{assessSD}} was applied. 
}
\usage{data(lw.sd.assessment)}
\format{A list.}


\keyword{datasets}
