\name{mas5.assessment}
\docType{instance}
\alias{mas5.assessment}
\title{An example of the result of an assessment}
\description{
  The Dilution and Spike-in cel files were processed with 
  MAS 5.0 software  and then the
  function \code{\link{assessAll}} was applied. 
}
\usage{data(rma.assessment)}
\format{A list of list.}


\keyword{datasets}
