\name{rma.assessment}
\docType{instance}
\alias{rma.assessment}
\title{An example of the result of an assessment}
\description{
  The Dilution and Spike-in cel files were processed with the
  \code{affy} version 1.0 package rma add-on function and then the
  function \code{\link{assessAll}} was applied. 
}
\usage{data(rma.assessment)}
\format{A list of list.}


\keyword{datasets}
