\name{hgu133a.spikein.xhyb}
\docType{instance}
\alias{hgu133a.spikein.xhyb}
\title{Cross hybridizers}
\description{Probe Sets likely to crosshybridize to spiked-in probesets in the
  Affymetrix HGU133A spike in}
This objact is list. Each component of the list contains probeset names
of possible crosshybridizers. The sequences of each spiked-in clone were
collected and blasted against all HG-U133A target sequences.  Target
sequences are the ~600bp regions from which probes were
selected. Thresholds of 100, 150 and 200bp were used and define the
three components of the list.
}
\usage{data(hgu133a.spikein.xhyb)}
\format{A list}
\source{Simon Cawley \verb+<simon_cawley@affymetrix.com>+}
\keyword{datasets}
