\name{mas5.assessment}
\docType{data}
\alias{mas5.assessment}
\alias{mas5.assessment.133}
\alias{mas5.assessment2}
\alias{mas5.assessment2.133}

\title{An example of the result of the assessments}
\description{
  The Dilution and both (HGU95 and HGU133) Spike-in cel files were
  processed with  
  MAS 5.0 software  and then the
  functions \code{\link{assessAll}} and \code{\link{assessSpikeIn2}}
  were applied. 
}
\usage{data(rma.assessment)
data(rma.assessment.133)
data(rma.assessment2)
data(rma.assessment2.133)
}
\format{A list of list.}


\keyword{datasets}
